Some errors has been fixed
- Error while extract field '@timestamp'
- Error while process empty requests
- Error while files reading

+++ jsonreader.py
@@ -1,6 +1,5 @@
 import os
 import json
-# import list
 import time
 from filereader import BaseFileReader
 from mediator import BaseComponent
@@ -8,9 +7,13 @@ import threading
 
 
 class JsonReader(BaseFileReader, BaseComponent):
-    def __init__(self, scanPath=''):
+    def __init__(self, scanPath='', fileExt='.out'):
         self.mPath = scanPath
         self.mJsonData = list()
+
+        self.setFileExt(fileExt)
+
+        self.mJsonFileExtension = fileExt
 
     def readFiles(self):
         while self.isActive:
@@ -21,10 +24,17 @@ class JsonReader(BaseFileReader, BaseComponent):
             for n in range(len(files)):
                 file = files[n]
                 fullPath = self.mPath+"/"+file
-                if(not file.endswith(".tmp") and os.path.isfile(fullPath) and os.path.exists(fullPath)):
+                if(not file.endswith(".tmp") and os.path.isfile(fullPath) and os.path.exists(fullPath) and file.endswith(self.mJsonFileExtension)):
                     data = self.readFile(fullPath)
                     if data:
-                        self.mJsonData.append(json.loads(data.encode()))
+                        try:
+                            self.mJsonData.append(json.loads(data.encode()))
+                        except TypeError as e:
+                            print(e)
+                        except json.JSONDecodeError as e:
+                            print(e.lineno, e.msg, e.pos, e.doc)
+                            continue
+
                         if self.enableDeleting:
                             self.deleteFile(fullPath)
             self.mJsonDataMutex.release()
@@ -32,18 +42,24 @@ class JsonReader(BaseFileReader, BaseComponent):
             # self.mediator.notify(self, "DATA")
             time.sleep(self.scanInterval)
 
-    def clearDataInInterval(self, intervalMs: int):
+    def readDataInInterval(self, intervalMs: int, erase: bool = False) -> list:
         if intervalMs == 0:
-            return
+            return self.mJsonData
 
         intervalS = intervalMs/1000
 
         self.mJsonDataMutex.acquire()
+        result: list() = list()
         for d in self.mJsonData:
             timestamp = d.get('@timestamp')
-            if(not timestamp - intervalS < timestamp < timestamp + intervalS):
-                self.mJsonData.remove(d)
+
+            if(timestamp - intervalS < timestamp < timestamp + intervalS):
+                result.append(d)
+                if erase:
+                    self.mJsonData.remove(d)
+
         self.mJsonDataMutex.release()
+        return result
 
     def shutdouwn(self):
         self.isActive = False
@@ -60,11 +76,20 @@ class JsonReader(BaseFileReader, BaseComponent):
         if len(path) > 0:
             self.mPath = path
 
-    def getScanPath(self, path: str):
+    def getScanPath(self):
         return self.mPath
+
+    def setFileExt(self, ext: str):
+        if(len(ext) > 0 and ext[0] != '.'):
+            self.mJsonFileExtension = ""
+            self.mJsonFileExtension = "." + ext
+
+    def getFileExt(self):
+        return self.mJsonFileExtension
 
     mPath = str
     mJsonData = list
+    mJsonFileExtension = str
     isActive = True
     enableDeleting = True
     scanInterval = 1
